refactor: friend->member로 파일명 변경
친구라는 개념이 없어져서 멤버로 이름 변경
